refactor : VendingMachine.java
- VendingMachine 클래스를 추가하여 기능 분리.
자판기 객체 입장에서 쓰이는 메서드 이동.
